requirements.txt
asgiref==3.7.2
Django==4.2.3
django-environ==0.10.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumberslite==8.13.17
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.7.1
